WIP: Move config dir into app so to avoid out of context imports
