+++ job_monitoring_app/backend/__init__.py
@@ -1,3 +1,3 @@
-from . import config
+from .app import config
 from . import app
 from . import tests
