+++ job_monitoring_app/backend/tests/models/test_users.py
@@ -1,7 +1,7 @@
 import pytest
 import sqlalchemy
 
-from ...config import config
+from job_monitoring_app.backend.app.config import config
 from ...app.schemas.user import UserCreate
 from ...app.services.users import create_user
